[reorganization]:Move series logic to its own app
Fixes #7351
