+++ specifyweb/urls.py
@@ -22,6 +22,7 @@ from specifyweb.backend.workbench import urls as wb_urls
 from specifyweb.backend.stats import urls as stat_urls
 from specifyweb.backend.merge import urls as merge_urls
 from specifyweb.backend.locality_update_tool import urls as locality_update_tool_urls
+from specifyweb.backend.series import urls as series_urls
 
 # print(get_resolver().reverse_dict.keys()) # Use for debugging urls
 
@@ -65,5 +66,6 @@ urlpatterns = [
     # url(r'^testcontext/', include()),
     path('stats/', include(stat_urls)),
     path('merge/', include(merge_urls)),
-    path('locality_update_tool/', include(locality_update_tool_urls))
+    path('locality_update_tool/', include(locality_update_tool_urls)),
+    path('series/', include(series_urls))
 ]
